+++ qt6_gui/models/settings.h
@@ -83,18 +83,18 @@ class Settings : public QObject
      * @brief Set FFT settings
      * @param aTransformSize New FFT size (must be power of 2)
      * @param aWindowType New window function type
-     * @throws std::invalid_argument if aTransformSize is zero
+     * @throws std::invalid_argument if aTransformSize is not positive
      *
      * These are set together because changing either requires recreating
      * FFT and window objects.
      */
-    void SetFFTSettings(size_t aTransformSize, FFTWindow::Type aWindowType);
+    void SetFFTSettings(int64_t aTransformSize, FFTWindow::Type aWindowType);
 
     /**
      * @brief Get the FFT size
      * @return Current FFT size
      */
-    [[nodiscard]] size_t GetFFTSize() const { return mFFTSize; }
+    [[nodiscard]] int64_t GetFFTSize() const { return mFFTSize; }
 
     /**
      * @brief Get the window function type
@@ -123,7 +123,10 @@ class Settings : public QObject
      * @brief Get the window stride based on current FFT size and window scale
      * @return Current window stride (FFT size / window scale)
      */
-    [[nodiscard]] size_t GetWindowStride() const { return mFFTSize / mWindowScale; }
+    [[nodiscard]] int64_t GetWindowStride() const
+    {
+        return mFFTSize / static_cast<int64_t>(mWindowScale);
+    }
 
     /**
      ** Display settings - Aperture (decibel range)
@@ -221,8 +224,8 @@ class Settings : public QObject
     void ApertureSettingsChanged();
 
   private:
-    static constexpr size_t KDefaultFFTSize = 2048;
-    size_t mFFTSize = KDefaultFFTSize;
+    static constexpr int64_t KDefaultFFTSize = 2048;
+    int64_t mFFTSize = KDefaultFFTSize;
     FFTWindow::Type mWindowType = FFTWindow::Type::Hann;
 
     // The window scale is the divisor applied to the FFT size to determine
